moved(FormModal): moved the component in the Modal folder
